+++ base.d.ts
@@ -2,6 +2,7 @@ export interface FieldConfig {
     type: any;
     immutable?: boolean;
     optional?: boolean;
+    required?: boolean;
     default?: any;
     enum?: any[];
     max?: number;
@@ -9,12 +10,13 @@ export interface FieldConfig {
     beforeChecks?: (value: any) => any;
     afterChecks?: (value: any) => any;
     validate?: (value: any) => void;
-    keys?: Record<string, FieldConfig>;
-    values?: FieldConfig;
+    keys?: Record<string, FieldConfig | Function>;
+    properties?: Record<string, FieldConfig | Function>;
+    values?: FieldConfig | Function;
     coerce?: boolean;
 }
 export interface SchemaDefinition {
-    [key: string]: FieldConfig;
+    [key: string]: Function | FieldConfig;
 }
 export interface parserConfig {
     safe?: boolean;
@@ -34,7 +36,7 @@ export interface errorObject {
 }
 export declare class ModelCoreError extends Error {
     source: string | Function | undefined;
-    path: Array<string | number> | string | undefined;
+    path: Array<string | number>;
     expected: any;
     received: any;
     code: string | undefined;
@@ -71,22 +73,37 @@ export declare class RequiredError extends ModelCoreError {
 export declare class ValueError extends ModelCoreError {
     static errorName: string;
 }
+export declare function Union<T extends readonly (abstract new (...args: any) => any)[]>(...args: T): {
+    new (): InstanceType<T[number]>;
+    unionTypes: T;
+    isArray(arg: any): arg is any[];
+    from<T_1>(arrayLike: ArrayLike<T_1>): T_1[];
+    from<T_1, U>(arrayLike: ArrayLike<T_1>, mapfn: (v: T_1, k: number) => U, thisArg?: any): U[];
+    from<T_1>(iterable: Iterable<T_1> | ArrayLike<T_1>): T_1[];
+    from<T_1, U>(iterable: Iterable<T_1> | ArrayLike<T_1>, mapfn: (v: T_1, k: number) => U, thisArg?: any): U[];
+    of<T_1>(...items: T_1[]): T_1[];
+    readonly [Symbol.species]: ArrayConstructor;
+};
 type UnwrapTypeConstructor<T> = T extends StringConstructor ? string : T extends NumberConstructor ? number : T extends BooleanConstructor ? boolean : T extends DateConstructor ? Date : T extends ArrayConstructor ? any[] : T extends ObjectConstructor ? Record<string, any> : T extends new (...args: any[]) => infer R ? R : unknown;
-type InferFieldRaw<T extends FieldConfig> = T['keys'] extends Record<string, FieldConfig> ? InferObject<T> : T['values'] extends FieldConfig ? InferArray<T> : UnwrapTypeConstructor<T['type']>;
-type InferField<T extends FieldConfig> = T['optional'] extends true ? InferFieldRaw<T> | undefined : InferFieldRaw<T>;
-type OptionalKeys<T extends Record<string, FieldConfig>> = {
-    [K in keyof T]: T[K]['optional'] extends true ? K : never;
+type NormalizeField<T> = T extends FieldConfig ? T : {
+    type: T;
+};
+type InferFieldRaw<T> = NormalizeField<T> extends infer F extends FieldConfig ? F['type'] extends {
+    unionTypes: readonly any[];
+} ? InstanceType<F['type']['unionTypes'][number]> : F['type'] extends ObjectConstructor ? InferObject<F> : F['type'] extends ArrayConstructor ? InferArray<F> : UnwrapTypeConstructor<F['type']> : any;
+type OptionalKeys<T extends Record<string, any>> = {
+    [K in keyof T]: NormalizeField<T[K]>['optional'] extends true ? K : NormalizeField<T[K]>['required'] extends false ? K : never;
 }[keyof T];
-type RequiredKeys<T extends Record<string, FieldConfig>> = {
-    [K in keyof T]: T[K]['optional'] extends true ? never : K;
+type RequiredKeys<T extends Record<string, any>> = {
+    [K in keyof T]: NormalizeField<T[K]>['optional'] extends true ? never : NormalizeField<T[K]>['required'] extends false ? never : K;
 }[keyof T];
-type InferObject<T extends FieldConfig> = T['keys'] extends Record<string, FieldConfig> ? {
+type InferObject<T extends FieldConfig> = T['keys'] extends Record<string, any> ? {
     [K in RequiredKeys<T['keys']>]: InferFieldRaw<T['keys'][K]>;
 } & {
     [K in OptionalKeys<T['keys']>]?: InferFieldRaw<T['keys'][K]>;
 } : Record<string, any>;
-type InferArray<T extends FieldConfig> = T['values'] extends FieldConfig ? Array<InferField<T['values']>> : any[];
-type SchemaToType<S extends SchemaDefinition> = {
+type InferArray<T extends FieldConfig> = T['type'] extends ArrayConstructor ? Array<InferFieldRaw<T['values']>> : any[];
+export type SchemaToType<S extends Record<string, any>> = {
     [K in RequiredKeys<S>]: InferFieldRaw<S[K]>;
 } & {
     [K in OptionalKeys<S>]?: InferFieldRaw<S[K]>;