+++ examples/inheritance.js
@@ -1,16 +1,15 @@
 import Base from "../base.js";
-
 class BaseModel extends Base {
     save() {
         // save to db
     }
 }
-
 class User extends BaseModel {
-    static collection = "users";
-    static schema = { name: { type: String } }
+}
+User.collection = "users";
+User.schema = {
+    name: { type: String }
 };
-
 const user = new User({ name: "John Doe" });
 user.save();
-console.log(user.name);
+console.log(user);